apps/ml-inference/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=9.5.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
albumentations>=1.3.0
timm>=0.9.0

pydantic>=2.0.0
kaggle>=1.5.16
roboflow>=1.1.0
wandb>=0.15.0
tensorboard>=2.13.0
efficientnet-pytorch>=0.7.1
pretrainedmodels>=0.7.4
psutil>=5.9.0
rich>=13.0.0
click>=8.1.0
python-dotenv>=1.0.0
